test: remove assumption about error message
It is unclear what changed, since this test started failing without any
code changes to strong-arc itself. Unfortunately it is unclear which
upstream dependency changed to cause this change in behaviour.

In the mean time, the rest of this test is still working, and does at
least test that the view selector changes the current view.